Update type definitions and class descriptions
Several modifications were carried out. The type 'dyadic_multiplier_t' is now defined as 'int64_t', enhancing its capacity. In 'scalar_array.h', headers were re-ordered, and class 'ScalarArray' description received a comprehensive update. New classes 'InterfaceBase' and 'ObjectBase' were introduced in 'device_object_base.h' with respective descriptions reflecting their usage across the codebase.

+++ core/include/roughpy/core/types.h
@@ -64,7 +64,7 @@ using param_t = double;
 using scalar_t = double;
 
 using resolution_t = int;
-using dyadic_multiplier_t = int;
+using dyadic_multiplier_t = int64_t;
 using dyadic_depth_t = resolution_t;
 
 using accuracy_t = param_t;
